Started work on a scripted traits editor
We can currently change the names and descriptions of scripted traits. The main window updates in response to these changes; however, it doesn't properly preserve the current selection in its navbar, if the user was viewing any scripted traits at the time; see (TODO) comment in main_window.cpp. We also need to implement adding, reordering, and deleting scripted traits.

Megalo game stats should now properly serialize string indices when their strings are changed (which can't happen yet).

Added ReachStringPicker::clearTarget, for use by containing UI when deselecting whatever data item has a string.

As a side note: these past few commits, I've been deleting comparison operators that I'd previously written for some data structures. This is largely because they're not as meaningful? Like, I've only just now figured out how to put it into words but basically, there needs to be a distinction between checking whether two things are *the same* object, and checking whether two things are *separate but identical* objects. JavaScript has operator === alongside operator == but unfortunately, C++ only has operator ==. Until I decide how I want to represent this distinction (and also create operators to check whether two ReachStrings are identical), it's not really possible to define *meaningful* operator==s for many of these classes.

+++ native/src/ReachVariantTool/ui/script_editor/page_script_options.cpp
@@ -125,6 +125,7 @@ ScriptEditorPageScriptOptions::ScriptEditorPageScriptOptions(QWidget* parent) : 
             return;
          }
          this->targetOption = list.emplace_back(new ReachMegaloOption);
+         this->targetOption->add_value(); // enum-options must have at least one value
          this->updateOptionFromVariant();
          this->updateOptionsListFromVariant();
          ReachEditorState::get().scriptOptionsModified();
@@ -231,6 +232,13 @@ ScriptEditorPageScriptOptions::ScriptEditorPageScriptOptions(QWidget* parent) : 
          if (index == 0) // can't move the first item up
             return;
          option.values.swap_items(index, index - 1);
+         {
+            auto c = option.currentValueIndex;
+            if (c == index)
+               c = index - 1;
+            else if (c == index - 1)
+               c = index;
+         }
          this->updateValuesListFromVariant();
          ReachEditorState::get().scriptOptionModified(this->targetOption);
       });
@@ -245,6 +253,13 @@ ScriptEditorPageScriptOptions::ScriptEditorPageScriptOptions(QWidget* parent) : 
          if (index == option.values.size() - 1) // can't move the last item down
             return;
          option.values.swap_items(index, index + 1);
+         {
+            auto c = option.currentValueIndex;
+            if (c == index)
+               c = index + 1;
+            else if (c == index + 1)
+               c = index;
+         }
          this->updateValuesListFromVariant();
          ReachEditorState::get().scriptOptionModified(this->targetOption);
       });
@@ -253,12 +268,17 @@ ScriptEditorPageScriptOptions::ScriptEditorPageScriptOptions(QWidget* parent) : 
             return;
          auto& option = *this->targetOption;
          auto  value  = this->targetValue;
-         auto  index  = option.values.index_of(value);
+         auto& list   = option.values;
+         if (list.size() == 1) {
+            QMessageBox::information(this, tr("Cannot remove option value"), tr("If a game variant contains enum-options with no values, that game variant won't be considered valid; it will not appear in the MCC's menus."));
+            return;
+         }
+         auto  index  = list.index_of(value);
          option.delete_value(value);
          if (index > 0)
-            this->targetValue = option.values[index - 1];
+            this->targetValue = list[index - 1];
          else if (option.values.size())
-            this->targetValue = option.values[0];
+            this->targetValue = list[0];
          else
             this->targetValue = nullptr;
          this->updateValueFromVariant();
@@ -334,6 +354,8 @@ void ScriptEditorPageScriptOptions::updateOptionFromVariant() {
    const QSignalBlocker blocker5(this->ui.optionType);
    //
    if (!this->targetOption) {
+      this->ui.optionName->clearTarget();
+      this->ui.optionDesc->clearTarget();
       //
       // TOOD: reset controls to a blank state
       //
@@ -375,6 +397,7 @@ void ScriptEditorPageScriptOptions::updateValuesListFromVariant() {
    }
    if (this->targetValue)
       _selectByPointerData(this->ui.listValues, this->targetValue);
+   this->ui.buttonValuesDelete->setDisabled(list.size() <= 1);
 }
 void ScriptEditorPageScriptOptions::updateValueFromVariant() {
    const QSignalBlocker blocker0(this->ui.valueName);
